requirements.txt
aiohttp==3.7.4.post0
aioschedule==0.5.2
aiosqlite==0.17.0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
click==8.0.4
colorama==0.4.4
idna==3.3
iso8601==0.1.16
multidict==5.2.0
nextcord==2.0.0a3
pydantic==1.8.2
pypika-tortoise==0.1.1
pytz==2021.3
tortoise-orm==0.17.8
typer==0.4.0
typing-extensions==3.10.0.2
yarl==1.7.2
